+++ internal/handlers/user/model.go
@@ -1,8 +1,8 @@
 package user
 
 type User struct {
-	ID       int    `json:"id"`
-	Login    string `json:"login"`
-	Password string `json:"password"`
-	Balance  int    `json:"balance"`
+	ID       int     `json:"id"`
+	Login    string  `json:"login"`
+	Password string  `json:"password"`
+	Balance  float64 `json:"balance"`
 }
